# public/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Space-Stylish E-commerce API
  description: |
    Space-Stylish E-commerce Website API Documentation

    This is a complete e-commerce platform API that provides the following features:
    - User registration, login and profile management
    - Product browsing, search and detailed information
    - Shopping cart and order management
    - Marketing campaign management
    - Sales reports and analytics
    - Admin functions

    ## Authentication
    Most APIs require JWT token authentication. Please include in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

    ## Rate Limiting
    - General APIs: 60 requests per second
    - API endpoints: 30 requests per second
    - Admin endpoints: 10 requests per second
    - Health check and static resources are excluded

    ## Response Format
    All APIs return a standardized response format:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message",
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

    ## Environment Variables
    - `API_VERSION`: API version (default: v1)
    - `SERVER_PORT`: Server port
    - `JWT_SIGN_SECRET`: JWT signing secret
    - `PARTNER_KEY`: TapPay partner key
    - `MERCHANT_ID`: TapPay merchant ID
    - `APP_ID`: Facebook App ID
    - `APP_SECRET`: Facebook App Secret

  version: 2.0.0
  contact:
    name: Space-Stylish Team
    email: support@space-stylish.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /api/v1/products:
    get:
      tags:
        - Products
      summary: Get products by category
      description: Get products with pagination and category filtering
      parameters:
        - name: category
          in: query
          description: Product category
          required: false
          schema:
            type: string
            enum: [all, women, men, accessories]
            default: all
        - name: paging
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
                properties:
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      pagination:
                        type: object
                        properties:
                          current:
                            type: integer
                          total:
                            type: integer
                          hasMore:
                            type: boolean
                          totalCount:
                            type: integer
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Products
      summary: Create product
      description: Create a new product (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - price
                - category
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                  enum: [women, men, accessories]
                texture:
                  type: string
                wash:
                  type: string
                place:
                  type: string
                note:
                  type: string
                story:
                  type: string
                main_image:
                  type: string
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      color_name:
                        type: string
                      color_code:
                        type: string
                      size:
                        type: string
                      stock:
                        type: integer
                images:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/products/search:
    get:
      tags:
        - Products
      summary: Search products
      description: Search products by keyword
      parameters:
        - name: keyword
          in: query
          description: Search keyword
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/products/{id}:
    get:
      tags:
        - Products
      summary: Get product detail
      description: Get detailed product information by ID
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Product'
                  message:
                    type: string
                    example: "Product detail retrieved successfully"
                  timestamp:
                    type: string
                    format: date-time
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Products
      summary: Update product
      description: Update product information (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                  enum: [women, men, accessories]
                texture:
                  type: string
                wash:
                  type: string
                place:
                  type: string
                note:
                  type: string
                story:
                  type: string
                main_image:
                  type: string
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      color_name:
                        type: string
                      color_code:
                        type: string
                      size:
                        type: string
                      stock:
                        type: integer
                images:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                  message:
                    type: string
                    example: "Product updated successfully"
                  timestamp:
                    type: string
                    format: date-time
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Product not found"
                  timestamp:
                    type: string
                    format: date-time
    delete:
      tags:
        - Products
      summary: Delete product
      description: Soft delete a product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                  message:
                    type: string
                    example: "Product deleted successfully"
                  timestamp:
                    type: string
                    format: date-time
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Product not found"
                  timestamp:
                    type: string
                    format: date-time
  /api/v1/products/{id}/variants:
    get:
      tags:
        - Products
      summary: Get product variants
      description: Get all variants of a specific product
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product variants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        color_code:
                          type: string
                        size:
                          type: string
                        stock:
                          type: integer
                  message:
                    type: string
                    example: "Product variants retrieved successfully"
                  timestamp:
                    type: string
                    format: date-time
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Product not found"
                  timestamp:
                    type: string
                    format: date-time

  # Marketing APIs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: "Error message"
        timestamp:
          type: string
          format: date-time
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          example: "Success message"
        timestamp:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        category:
          type: string
          enum: [women, men, accessories]
          example: "women"
        title:
          type: string
          example: "Summer Dress"
        description:
          type: string
          example: "Beautiful summer dress"
        price:
          type: number
          example: 1000
        texture:
          type: string
          example: "Cotton"
        wash:
          type: string
          example: "Machine wash"
        place:
          type: string
          example: "Taiwan"
        note:
          type: string
          example: "Note about the product"
        story:
          type: string
          example: "Product story"
        main_image:
          type: string
          example: "/images/product1.jpg"
        colors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: "#FF0000"
              name:
                type: string
                example: "Red"
        sizes:
          type: array
          items:
            type: string
          example: ["S", "M", "L"]
        variants:
          type: array
          items:
            type: object
            properties:
              color_name:
                type: string
              color_code:
                type: string
              size:
                type: string
              stock:
                type: integer
        images:
          type: array
          items:
            type: string
...